Move filter models to shared dir
