+++ backend/tests/unit/filter/filter.controller.spec.ts
@@ -2,11 +2,11 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { FilterController } from '../../../src/filter/filter.controller';
+import { FilterService } from '../../../src/filter/filter.service';
 import {
   EdgeTypeFilterModel,
-  FilterService,
   NodeTypeFilterModel,
-} from '../../../src/filter/filter.service';
+} from '../../../src/shared/filter';
 import { FilterServiceBase } from '../../../src/filter/filter.service.base';
 import {
   getEdgeTypeFilterModelResult,
