+++ shared/src/filter/EdgeTypeFilterModel.ts
@@ -1,0 +1,6 @@
+import FilterModelEntry from './FilterModelEntry';
+
+export default interface EdgeTypeFilterModel {
+  name: string;
+  properties: FilterModelEntry[];
+}
