+++ shared/src/filter/FilterModelEntry.ts
@@ -1,0 +1,6 @@
+import { Property } from '../entities/Property';
+
+export default interface FilterModelEntry {
+  key: string;
+  values: Property[];
+}
