+++ shared/src/filter/NodeTypeFilterModel.ts
@@ -1,0 +1,6 @@
+import FilterModelEntry from './FilterModelEntry';
+
+export default interface NodeTypeFilterModel {
+  name: string;
+  properties: FilterModelEntry[];
+}
