+++ shared/src/filter/index.ts
@@ -1,0 +1,5 @@
+import EdgeTypeFilterModel from './EdgeTypeFilterModel';
+import FilterModelEntry from './FilterModelEntry';
+import NodeTypeFilterModel from './NodeTypeFilterModel';
+
+export { EdgeTypeFilterModel, FilterModelEntry, NodeTypeFilterModel };
